Build changes: esbuild & TypeScript
Aim is to allow the majority of the JavaScript to move to TypeScript and
use multiple files as required while continuing to provide it as a
single JavaScript file for emscripten.

As a bonus we can clean up the SVG loading by bundling it, removing
async awkwardness at initialization time.