Work towards building on the audio-recording branch
On the basis that's as close as we can get to the beta for now.
Still has HAL issues.

One local change to MicroPython was needed to include mphal.h in
modspeech - need to review how this works for the non-sim build as it
seems like it should be included.